Make EntityRef a friend of Entity
Both are EntityBase.

close #23
